+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 - Send entity operations events to all registered plugins
+- Add type filter to get services api
 ### Changed
+- Allow plugin users to create and get operator users
 ### Fixed
+- Fix role-based permissions in get user api
 ### Removed
 
 ## [1.0.0-alpha.8] - 2018-11-13